// apps/desktop/src/lib/generated/GradeReviewRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GradeDto } from "./GradeDto";

export type GradeReviewRequest = { review_event_id: string, card_id: string, card_content_version: number, schedule_version: number, raw_response: string, chosen_grade: GradeDto, response_duration_ms: number, };